Slight changes to cameras, start of GameWorld class, import of other classes
Cameras are slightly different, and now allow for setting a min/max yaw
and pitch. This includes new mutators to change these min/max variables.
There's the start of a GameWorld class, which has static members
containing a vector of GameObjects, a world octree, and a world gravity.
The GameWorld class isn't used at all yet, but will eventually run the
main GameObject update loop. GameObjects will be able to reference the
GameWorld's static octree/gravity variables for physics and collision
calculations.

Many new classes have been imported from other projects, including
Quaternions, OBBs, and the aforementioned ColliderTree (implements
octrees), which uses a Node class. I slightly rewrote the
ColliderTree/Node classes to use OBBs instead of the basic Collider
class used before which acted more like an AABB. Also moved/renamed
stuff so that the OBB has a collidesWith method that accepts a reference
to another OBB and returns whether a collision happened, using the
separating axis theorem. It may be good, in the future, to return the
collider we hit, for use in physics, but that's probably not necessary
for the scope of this project.

Most of the actual currently used code is exactly the same and the
program doesn't do anything different at the moment - it still just
renders a spinning, lit, textured Cthulhu with camera controls.